pyproject.toml
[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = ""
authors = ["Jang JiWoo <77304817+cheeou@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
databases = "^0.9.0"
pydantic = {extras = ["email"], version = "^2.10.4"}
bcrypt = "^4.2.1"
python-jose = "^3.3.0"
python-jwt = "^4.1.0"
argon2-cffi = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
